+++ src/javascript/binary/static_pages/home.js
@@ -2,8 +2,10 @@ var submit_email = require('../websocket_pages/user/verify_email').submit_email;
 
 var Home = (function() {
     var init = function() {
-        check_login_hide_signup();
-        submit_email();
+        if (!page.client.redirect_if_login()) {
+            check_login_hide_signup();
+            submit_email();
+        }
     };
     var check_login_hide_signup = function() {
         if (page.client.is_logged_in) {
