+++ src/javascript/binary/static_pages/static_pjax.js
@@ -16,9 +16,7 @@ var CharityPage = require('./charity').CharityPage;
 pjax_config_page('/home', function() {
     return {
         onLoad: function() {
-            if(!page.client.redirect_if_login()) {
-                Home.init();
-            }
+            Home.init();
         }
     };
 });
@@ -29,7 +27,7 @@ pjax_config_page('/why-us', function() {
             WhyUs.init();
         },
         onUnload: function() {
-            $(window).off('scroll');
+            Scroll.offScroll();
         }
     };
 });
@@ -37,12 +35,10 @@ pjax_config_page('/why-us', function() {
 pjax_config_page('/volidx-markets', function() {
     return {
         onLoad: function() {
-            if (page.url.location.hash !== "") {
-              $('a[href="' + page.url.location.hash + '"]').click();
-            }
+            Scroll.goToHashSection();
         },
         onUnload: function() {
-            $(window).off('scroll');
+            Scroll.offScroll();
         }
     };
 });
@@ -53,7 +49,7 @@ pjax_config_page('/open-source-projects', function() {
             Scroll.sidebar_scroll($('.open-source-projects'));
         },
         onUnload: function() {
-            $(window).off('scroll');
+            Scroll.offScroll();
         }
     };
 });
@@ -64,7 +60,7 @@ pjax_config_page('/payment-agent', function() {
             Scroll.sidebar_scroll($('.payment-agent'));
         },
         onUnload: function() {
-            $(window).off('scroll');
+            Scroll.offScroll();
         }
     };
 });
@@ -72,12 +68,10 @@ pjax_config_page('/payment-agent', function() {
 pjax_config_page('/get-started', function() {
     return {
         onLoad: function() {
-            if (!/get-started-jp/.test(window.location.pathname)) {
-                GetStarted.get_started_behaviour();
-            }
+            GetStarted.get_started_behaviour();
         },
         onUnload: function() {
-            $(window).off('scroll');
+            Scroll.offScroll();
         },
     };
 });
@@ -109,25 +103,8 @@ pjax_config_page('/charity', function() {
 pjax_config_page('/terms-and-conditions', function() {
     return {
         onLoad: function() {
-            var selected_tab = page.url.params_hash().selected_tab;
-            if(selected_tab) {
-              $('li#' + selected_tab + ' a').click();
-            }
-            var year = document.getElementsByClassName('currentYear');
-            for (i = 0; i < year.length; i++){
-              year[i].innerHTML = new Date().getFullYear();
-            }
+            TermsAndConditions.init();
         },
-    };
-});
-
-pjax_config_page('\/login|\/loginid_switch', function() {
-    return {
-        onLoad: function() {
-            if(!$('body').hasClass('BlueTopBack')) {
-                window.location.href = Login.login_url();
-            }
-        }
     };
 });
 
@@ -201,13 +178,7 @@ pjax_config_page('/get-started-jp', function() {
 
 pjax_config_page('/open-positions', function() {
   return {
-      onLoad: function() {
-        if (document.getElementById('Information_Technology')) {
-          if (page.url.location.hash) {
-              $.scrollTo($(page.url.location.hash));
-          }
-        }
-      }
+      onLoad: function() {if (/\/open-positions\.html/.test(window.location.pathname)) Scroll.scrollToHashSection();}
   };
 });
