+++ src/javascript/binary/websocket_pages/cashier/account_transferws.js
@@ -251,9 +251,25 @@ var account_transferws = (function(){
 
     };
 
+    var onLoad = function() {
+        BinarySocket.init({
+            onmessage: function(msg){
+                var response = JSON.parse(msg.data);
+                if (response) {
+                    account_transferws.apiResponse(response);
+                }
+            }
+        });
+
+        if(TUser.get().hasOwnProperty('is_virtual')) {
+            account_transferws.init();
+        }
+    };
+
     return {
         init : init,
-        apiResponse : apiResponse
+        apiResponse : apiResponse,
+        onLoad: onLoad,
     };
 
 })();
