+++ src/javascript/binary/websocket_pages/cashier/cashier.js
@@ -1,3 +1,5 @@
+var japanese_client = require('../../common_functions/country_base').japanese_client;
+
 var Cashier = (function() {
     "use strict";
 
@@ -41,9 +43,32 @@ var Cashier = (function() {
         $a.replaceWith($('<a/>', {class: $a.attr('class').replace('pjaxload') + ' button-disabled', html: $a.html()}));
     };
 
+    var onLoad = function() {
+        if (!/\/cashier\.html/.test(window.location.pathname) || !page.client.is_logged_in) {
+            return;
+        } else {
+            Cashier.check_locked();
+            Cashier.check_virtual_top_up();
+            page.contents.topbar_message_visibility(TUser.get().landing_company);
+        }
+    };
+
+    var onLoadPaymentMethods = function() {
+        if (japanese_client()) {
+            window.location.href = page.url.url_for('/');
+        }
+        if (!page.client.is_logged_in || page.client.is_virtual()) {
+            return;
+        } else {
+            Cashier.check_locked();
+        }
+    };
+
     return {
         check_locked: check_locked,
-        check_virtual_top_up: check_virtual_top_up
+        check_virtual_top_up: check_virtual_top_up,
+        onLoad: onLoad,
+        onLoadPaymentMethods: onLoadPaymentMethods,
     };
 }());
