+++ src/javascript/binary/websocket_pages/trade/message.js
@@ -1,5 +1,6 @@
 var PortfolioWS = require('../user/account/portfolio/portfolio.init').PortfolioWS;
 var ProfitTableWS = require('../user/account/profit_table/profit_table.init').ProfitTableWS;
+var StatementWS = require('../user/account/statement/statement.init').StatementWS;
 
 /*
  * This Message object process the response from server and fire
