+++ src/javascript/binary/websocket_pages/user/account/statement/statement.ui.js
@@ -4,6 +4,7 @@ var Button = require('../../../../common_functions/attach_dom/button').Button;
 var Table = require('../../../../common_functions/attach_dom/table').Table;
 var showTooltip = require('../../../../common_functions/get_app_details').showTooltip;
 var japanese_client = require('../../../../common_functions/country_base').japanese_client;
+var Statement = require('../statement').Statement;
 
 var StatementUI = (function(){
     "use strict";
@@ -107,9 +108,7 @@ var StatementUI = (function(){
         updateStatementTable: updateStatementTable,
         errorMessage: errorMessage,
         exportCSV: exportCSV,
-        setOauthApps: function(values) {
-            return (oauth_apps = values);
-        }
+        setOauthApps: function(values) { return (oauth_apps = values); },
     };
 }());
