+++ src/javascript/binary/websocket_pages/websocket_pjax.js
@@ -8,7 +8,6 @@ var MarketTimesUI = require('./resources/market_times/market_timesws.ui').Market
 var AuthenticateWS = require('./user/account/authenticate').AuthenticateWS;
 var PasswordWS = require('./user/account/change_password').PasswordWS;
 var PaymentAgentTransferSocket = require('./user/account/payment_agent_transfer').PaymentAgentTransferSocket;
-var japanese_client = require('../common_functions/country_base').japanese_client;
 var PortfolioWS = require('./user/account/portfolio/portfolio.init').PortfolioWS;
 var ProfitTableWS = require('./user/account/profit_table/profit_table.init').ProfitTableWS;
 var APITokenWS = require('./user/account/settings/api_token').APITokenWS;
@@ -19,6 +18,8 @@ var Limits = require('./user/account/settings/limits').Limits;
 var SelfExclusionWS = require('./user/account/settings/self_exclusion').SelfExclusionWS;
 var SettingsDetailsWS = require('./user/account/settings/settings_detailsws').SettingsDetailsWS;
 var SecurityWS = require('./user/account/settings/settings_securityws').SecurityWS;
+var SettingsWS = require('./user/account/settings').SettingsWS;
+var StatementWS = require('./user/account/statement/statement.init').StatementWS;
 
 pjax_config_page_require_auth("user/profit_table", function(){
     return {
@@ -35,7 +36,7 @@ pjax_config_page_require_auth("user/statement", function(){
     return {
         onLoad: function() {
             StatementWS.init();
-            Statement.attachDatePicker();
+            StatementWS.attachDatePicker();
         },
         onUnload: function() {
             StatementWS.clean();
@@ -97,18 +98,7 @@ pjax_config_page_require_auth("user/security/cashier_passwordws", function() {
 pjax_config_page_require_auth("account/account_transferws", function() {
     return {
         onLoad: function() {
-            BinarySocket.init({
-                onmessage: function(msg){
-                    var response = JSON.parse(msg.data);
-                    if (response) {
-                        account_transferws.apiResponse(response);
-                    }
-                }
-            });
-
-            if(TUser.get().hasOwnProperty('is_virtual')) {
-                account_transferws.init();
-            }
+            account_transferws.onLoad();
         }
     };
 });
@@ -116,13 +106,7 @@ pjax_config_page_require_auth("account/account_transferws", function() {
 pjax_config_page("/cashier", function(){
     return {
         onLoad: function() {
-          if (!/\/cashier\.html/.test(window.location.pathname) || !page.client.is_logged_in) {
-              return;
-          } else {
-              Cashier.check_locked();
-              Cashier.check_virtual_top_up();
-              page.contents.topbar_message_visibility(TUser.get().landing_company);
-          }
+            Cashier.onLoad();
         }
     };
 });
@@ -130,14 +114,7 @@ pjax_config_page("/cashier", function(){
 pjax_config_page("/cashier/payment_methods", function(){
     return {
         onLoad: function() {
-            if (japanese_client()) {
-                window.location.href = page.url.url_for('/');
-            }
-            if (!page.client.is_logged_in || page.client.is_virtual()) {
-                return;
-            } else {
-                Cashier.check_locked();
-            }
+            Cashier.onLoadPaymentMethods();
         }
     };
 });
@@ -145,7 +122,7 @@ pjax_config_page("/cashier/payment_methods", function(){
 pjax_config_page_require_auth("cashier/forwardws|cashier/epg_forwardws", function() {
     return {
         onLoad: function() {
-          ForwardWS.checkOnLoad();
+            ForwardWS.checkOnLoad();
         }
     };
 });
@@ -153,18 +130,7 @@ pjax_config_page_require_auth("cashier/forwardws|cashier/epg_forwardws", functio
 pjax_config_page("payment_agent_listws", function() {
     return {
         onLoad: function() {
-            BinarySocket.init({
-                onmessage: function(msg) {
-                    var response = JSON.parse(msg.data);
-                    if (response) {
-                        if (response.msg_type === "paymentagent_list") {
-                            PaymentAgentListWS.responseHandler(response);
-                        }
-                    }
-                }
-            });
-            Content.populate();
-            PaymentAgentListWS.init();
+            PaymentAgentListWS.onLoad();
         }
     };
 });
@@ -246,6 +212,14 @@ pjax_config_page_require_auth("settings/detailsws", function() {
     return {
         onLoad: function() {
             SettingsDetailsWS.onLoad();
+        }
+    };
+});
+
+pjax_config_page_require_auth("settingsws|securityws", function() {
+    return {
+        onLoad: function() {
+            SettingsWS.onLoad();
         }
     };
 });
